move: IndexBookResponse 위치변경
