+++ src/main/java/com/nhnacademy/store99/front/book/service/BookService.java
@@ -2,8 +2,8 @@ package com.nhnacademy.store99.front.book.service;
 
 import com.nhnacademy.store99.front.book.dto.response.BookPageResponse;
 import com.nhnacademy.store99.front.book.dto.response.BookResponse;
-import com.nhnacademy.store99.front.book.dto.response.IndexBookResponse;
 import com.nhnacademy.store99.front.common.response.CustomPageImpl;
+import com.nhnacademy.store99.front.index.dto.response.IndexBookResponse;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
