+++ src/main/java/com/nhnacademy/store99/front/book/service/BookServiceImpl.java
@@ -4,9 +4,9 @@ package com.nhnacademy.store99.front.book.service;
 import com.nhnacademy.store99.front.book.adapter.BookAdapter;
 import com.nhnacademy.store99.front.book.dto.response.BookPageResponse;
 import com.nhnacademy.store99.front.book.dto.response.BookResponse;
-import com.nhnacademy.store99.front.book.dto.response.IndexBookResponse;
 import com.nhnacademy.store99.front.common.response.CommonResponse;
 import com.nhnacademy.store99.front.common.response.CustomPageImpl;
+import com.nhnacademy.store99.front.index.dto.response.IndexBookResponse;
 import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
